removed additional js files
